fast_api/requirements.txt
fastapi
uvicorn
nltk==3.9b1
openai
llama-index
llama-index-llms-openai
pypdf
pydantic-settings
